requirements.txt
streamlit
wikipedia-api
requests
beautifulsoup4
fpdf
PyPDF2
langchain
python-dotenv
